# portal-api-develop/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "iemanager",
    "description": "IEManger Application Backend APIs",
    "contact": {}
  },
  "host": "",
  "basePath": "/iemanager",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/subscription-packages": {
      "get": {
        "description": "List all avaliable packages in FA system",
        "summary": "Get subscription packages",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiV1SubscriptionPackagesGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/api/v1/user-subscription-packages": {
      "get": {
        "description": "List all Subscriptin pakcages by user",
        "summary": "Get user subscribed packages",
        "tags": [
          "IEManager"
        ],
        "operationId": "ApiV1UsersUserIdSubscriptionPackagesGet",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "account_id",
            "in": "header",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "Customer",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
...
}